Refactor contact section components
