+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import { GetStaticProps } from "next"
 import Head from "next/head"
 import styled from "styled-components"
 import Blogs from "../components/Blogs"
-import ContactSection from "../components/ContactSection"
+import ContactSection from "../components/contact/ContactSection"
 import Footer from "../components/Footer"
 import NavBar from "../components/navbar/NavBar"
 import Projects from "../components/Projects"
